requirements.txt
fastapi[standard]
requests
python-dotenv
pyttsx3
openai-whisper
python-multipart
Faker
pytest-html